feat: support Windows releases
