+++ README.md
@@ -140,7 +140,8 @@ turso db destroy <database name>
 The `turso` program keeps settings in your local machine in the following base
 directory in `turso/settings.json` file:
 
-| OS    | Config directory                          |
-| ----- | ----------------------------------------- |
-| Linux | `$XDG_CONFIG_HOME` or `$HOME/.config`     |
-| macOS | `$HOME/Library/Application Support/turso` |
+| OS      | Config directory                          |
+| ------- | ----------------------------------------- |
+| Linux   | `$XDG_CONFIG_HOME` or `$HOME/.config`     |
+| macOS   | `$HOME/Library/Application Support/turso` |
+| Windows | `%APPDATA%\turso`                         |
