+++ internal/cmd/update.go
@@ -92,21 +92,3 @@ var updateCmd = &cobra.Command{
 		return Update()
 	},
 }
-
-func Update() error {
-	var updateCmd string
-
-	if IsUnderHomebrew() {
-		updateCmd = "brew update && brew upgrade turso"
-	} else {
-		updateCmd = "curl -sSfL \"https://get.tur.so/install.sh\" | sh"
-	}
-	command := exec.Command("sh", "-c", updateCmd)
-	command.Stdout = os.Stdout
-	command.Stderr = os.Stderr
-	err := command.Run()
-	if err != nil {
-		return fmt.Errorf("failed to execute update command: %w", err)
-	}
-	return nil
-}
